requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
datasets==3.2.0
dill==0.3.8
diskcache==5.6.3
einops==0.8.0
evaluate==0.4.3
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
llama_cpp_python==0.3.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
optimum==1.23.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.1.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
tree-sitter==0.24.0
tree-sitter-typescript==0.23.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
